Create autonotifynationalsecurityahencyofexploitorintrusion.py
In this script, the notify_national_security_agency function takes the event type and details as input. It prepares the notification message and sends it to the specified URL (replace it with the actual URL provided by the national security agency). The function uses the requests library to make an HTTP POST request to the notification endpoint.

In the main section of the script, the user is prompted to enter the event type and details. These could be obtained from your intrusion detection system or other security monitoring tools. Then, the notify_national_security_agency function is called with the provided event type and details to send the notification.

Please note that this script is a general example, and you should adapt it to the specific requirements and guidelines provided by the national security agency you intend to notify.