Create autoblacklistvmware.py
In this script, we create a class VMwareBlacklist that represents the VMware IP blacklist. It has a list attribute blacklist to store the blacklisted IP addresses.

The add_to_blacklist method allows adding IP addresses to the blacklist by appending them to the blacklist list. The remove_from_blacklist method allows removing IP addresses from the blacklist if they exist in the list.

The update_vmware_blacklist method runs the command to update the VMware blacklist using the vmware-cli command-line tool. It iterates over the blacklist list and adds each IP address to the command. The subprocess.run function is used to execute the command.

In the main section of the script, an instance of the VMwareBlacklist class is created. IP addresses are added to the blacklist using the add_to_blacklist method, and an IP address is removed using the remove_from_blacklist method. Finally, the update_vmware_blacklist method is called to update the VMware blacklist.

Please note that the script assumes the existence of the vmware-cli command-line tool for managing the VMware blacklist. You may need to adjust the command or install additional dependencies based on your specific VMware setup and requirements.