Create autocreatemodule.py
This script prompts the user to enter the name of the module they want to create. It creates a directory with the lowercase version of the module name and creates a Python file inside it with the same name as the module. The module file contains intentional content, such as a hello() function that prints a greeting message.

You can run this script using Python, enter the desired module name when prompted, and it will create the module with the intentional content in the specified directory.